+++ packages/tools/generator-widget/generators/app/index.js
@@ -236,22 +236,12 @@ class MxGenerator extends Generator {
     _writeEndToEndTests() {
         if (this.widget.hasE2eTests && this.widget.isPlatformWeb) {
             if (this.widget.isLanguageTS) {
-                this._copyFile(
-                    `${this.widget.templateSourcePath}tests/configs/e2e/tsconfig.json`,
-                    "configs/e2e/tsconfig.json"
-                );
+                this._copyFile(`${this.widget.templateSourcePath}tests/tsconfig.json`, "cypress/tsconfig.json");
             }
 
             this._copyTemplate(
                 `${this.widget.templateSourcePath}tests/e2e/WidgetName.spec.${this.widget.fileExtension}.ejs`,
-                `cypress/integration/${this.widget.name}.spec.${this.widget.fileExtension}`
-            );
-
-            this._copyFile(`${this.widget.templateSourcePath}tests/e2e/support/e2e.js`, `cypress/support/e2e.js`);
-
-            this._copyFile(
-                `${this.widget.templateSourcePath}configs/e2e/cypress/support/e2e.js`,
-                `configs/e2e/cypress/support/e2e.js`
+                `cypress/e2e/${this.widget.name}.spec.${this.widget.fileExtension}`
             );
         }
     }