+++ packages/tools/generator-widget/generators/app/templates/pluggable/web/emptyTemplateJs/tests/e2e/support/e2e.js
@@ -1,2 +1,0 @@
-import "../configs/e2e/cypress/support/command";
-