+++ packages/tools/generator-widget/generators/app/index.js
@@ -236,17 +236,12 @@ class MxGenerator extends Generator {
     _writeEndToEndTests() {
         if (this.widget.hasE2eTests && this.widget.isPlatformWeb) {
             if (this.widget.isLanguageTS) {
-                this._copyFile(`${this.widget.templateSourcePath}tests/e2e/tsconfig.json`, "tests/e2e/tsconfig.json");
+                this._copyFile(`${this.widget.templateSourcePath}tests/tsconfig.json`, "cypress/tsconfig.json");
             }
 
             this._copyTemplate(
                 `${this.widget.templateSourcePath}tests/e2e/WidgetName.spec.${this.widget.fileExtension}.ejs`,
-                `tests/e2e/${this.widget.name}.spec.${this.widget.fileExtension}`
-            );
-
-            this._copyFile(
-                `${this.widget.templateSourcePath}tests/e2e/pages/home.page.${this.widget.fileExtension}.ejs`,
-                `tests/e2e/pages/home.page.${this.widget.fileExtension}`
+                `cypress/e2e/${this.widget.name}.spec.${this.widget.fileExtension}`
             );
         }
     }